+++ src/client/terrain/systems.rs
@@ -1,14 +1,9 @@
 use crate::prelude::*;
 
 pub fn prepare_spawn_area_system(mut client: ResMut<RenetClient>) {
-    let render_distance = 2;
-
     info!("Sending chunk requests for spawn area");
 
-    let chunks = terrain_resources::ChunkManager::instantiate_chunks(
-        Vec3::new(0.0, 0.0, 0.0),
-        render_distance,
-    );
+    let chunks = ChunkManager::instantiate_chunks(Vec3::ZERO, Vec3::ONE);
 
     let positions: Vec<Vec3> = chunks.into_iter().map(|chunk| chunk.position).collect();
     let message = bincode::serialize(&NetworkingMessage::ChunkBatchRequest(positions));
@@ -18,9 +13,9 @@ pub fn prepare_spawn_area_system(mut client: ResMut<RenetClient>) {
 
 pub fn generate_world_system(
     mut client: ResMut<RenetClient>,
-    mut chunk_manager: ResMut<terrain_resources::ChunkManager>,
+    mut chunk_manager: ResMut<ChunkManager>,
 ) {
-    let render_distance = 6;
+    let render_distance = Vec3::new(4.0, 4.0, 4.0);
 
     info!("Sending chunk requests for chunks");
 
@@ -28,16 +23,17 @@ pub fn generate_world_system(
 
     let positions: Vec<Vec3> = chunks.into_iter().map(|chunk| chunk.position).collect();
 
-    let batched_positions = positions.chunks(32);
+    let batched_positions = positions.chunks(16);
+    assert!(batched_positions.len() > 0, "Batched positions is empty");
 
-    batched_positions.for_each(|batch| {
+    batched_positions.enumerate().for_each(|(index, batch)| {
         let request_positions = batch.to_vec();
         info!(
             "Sending chunk batch request for {:?}",
             request_positions.len()
         );
         let message = bincode::serialize(&NetworkingMessage::ChunkBatchRequest(request_positions));
-        info!("requesting chunks");
+        info!("requesting chunks #{}", index);
         client.send_message(DefaultChannel::ReliableUnordered, message.unwrap());
     });
 }
@@ -48,7 +44,7 @@ pub fn handle_chunk_mesh_update_events(
     asset_server: Res<AssetServer>,
     mut meshes: ResMut<Assets<Mesh>>,
     mut materials: ResMut<Assets<StandardMaterial>>,
-    mut chunk_manager: ResMut<terrain_resources::ChunkManager>,
+    chunk_manager: ResMut<ChunkManager>,
     mut chunk_mesh_update_events: EventReader<terrain_events::ChunkMeshUpdateEvent>,
     mut mesh_query: Query<(Entity, &terrain_components::ChunkMesh)>,
     texture_manager: ResMut<terrain_util::TextureManager>,
@@ -62,7 +58,7 @@ pub fn handle_chunk_mesh_update_events(
         match chunk_option {
             Some(chunk) => {
                 for (entity, chunk_mesh) in mesh_query.iter_mut() {
-                    if lib::Chunk::key_eq_pos(chunk_mesh.key, chunk.position) {
+                    if Chunk::key_eq_pos(chunk_mesh.key, chunk.position) {
                         commands.entity(entity).despawn();
                     }
                 }
@@ -87,7 +83,7 @@ fn add_chunk_objects(
     asset_server: &Res<AssetServer>,
     meshes: &mut ResMut<Assets<Mesh>>,
     materials: &mut ResMut<Assets<StandardMaterial>>,
-    chunk: &lib::Chunk,
+    chunk: &Chunk,
     texture_manager: &terrain_util::TextureManager,
 ) {
     if let Some(mesh) = create_chunk_mesh(chunk, texture_manager) {
@@ -104,7 +100,7 @@ fn add_chunk_objects(
 }
 
 fn create_chunk_mesh(
-    chunk: &lib::Chunk,
+    chunk: &Chunk,
     texture_manager: &terrain_util::TextureManager,
 ) -> Option<Mesh> {
     terrain_util::create_chunk_mesh(chunk, texture_manager)
@@ -146,7 +142,7 @@ fn spawn_chunk(
     meshes: &mut Mut<Assets<Mesh>>,
     material: Handle<StandardMaterial>,
     mesh: Mesh,
-    chunk: &lib::Chunk,
+    chunk: &Chunk,
 ) {
     commands.spawn((
         Mesh3d(meshes.add(mesh)),