+++ src/server/prelude.rs
@@ -19,16 +19,19 @@ pub use bevy_renet::netcode::ServerConfig;
 pub use bevy_renet::RenetServerPlugin;
 pub use renet::DefaultChannel;
 pub use renet::*;
+pub use serde::Deserialize;
 
-pub use self::lib::Chunk;
-pub use self::terrain_util::Block;
+// other crates
+pub use rayon::iter::IntoParallelIterator;
+pub use rayon::iter::IntoParallelRefMutIterator;
+pub use rayon::iter::ParallelIterator;
+
+pub use lib::*;
 pub use noise::NoiseFn;
 pub use noise::Perlin;
+pub use rsmc as lib;
 
 // my crates
-pub use rsmc as lib;
-pub use rsmc::BlockId;
-
 pub use crate::networking::systems as networking_systems;
 
 pub use crate::player::resources as player_resources;