+++ src/server/terrain/systems.rs
@@ -1,19 +1,13 @@
-use rayon::iter::IntoParallelRefMutIterator;
-use rayon::iter::ParallelIterator;
-
 use crate::prelude::*;
 
-pub fn setup_world_system(mut chunk_manager: ResMut<terrain_resources::ChunkManager>) {
+pub fn setup_world_system(mut chunk_manager: ResMut<ChunkManager>) {
     let generator = terrain_util::generator::Generator::new(0);
 
-    let render_distance = 16;
+    let render_distance = Vec3::new(12.0, 2.0, 12.0);
 
     info!("Generating chunks");
 
-    let mut chunks = terrain_resources::ChunkManager::instantiate_chunks(
-        Vec3::new(0.0, 0.0, 0.0),
-        render_distance,
-    );
+    let mut chunks = ChunkManager::instantiate_chunks(Vec3::ZERO, render_distance);
 
     chunks.par_iter_mut().for_each(|chunk| {
         info!("Generating chunk at {:?}", chunk.position);