+++ src/server/terrain/util/generator.rs
@@ -1,5 +1,4 @@
 use crate::prelude::*;
-use lib::CHUNK_SIZE;
 
 pub struct Generator {
     pub seed: u32,